Move api/v2 protobuf to root
Separate the api protobuf definition from the v2 implementation. The
intent is to make the protobuf be implementation agnostic and use `buf`
tooling to generate code in an internal package.

- use buf remote plugins; don't need `tools.go` to add the protoc plugins
- generated code is placed in v2/internal/api/v2

Will try to further remove annotations from pi.proto as buf tooling
allows in the future; for now can't use managed mode.