Chore: update pre-commit and release actions
More vacillation; coming to the conclusion that testing of changes
should happen in the pre-commit action (push to main and pr) only. The
release action should only build the assets and trigger a Go Report
update. That means that I need to resist the urge to tag locally
until I know all the actions have passed.

- Add golangci-lint back to pre-commit GitHub action, but in a separate
  runner instance to avoid collision with go tests
- Remove linter from release, but do trigger Go Report Card refresh on
  successful generation of assets