feat(collate): Add a client command to print pi
For most use cases, the client is used to generate requests and the
results are largely irrelevant. This feature adds a client subcommand
`collate` that creates a target buffer and adds resulting digits to it
as results are returned. This allows using the base client when collating
the results are unneeded.