fix: Prefer int64 over uint64 to reduce casting
It was probably a big mistake to use uints in v2 protobuf declaration,
as most of the math requires ints and OTEL for go only supports ints.

This change deals with the uint/int conversion in FetchDigit functions
and will return an error if a uint index is provided that can't be cast
as an int without overflowing.